@@ -6,7 +6,7 @@
 
 	<groupId>org.aksw.simba.lsq</groupId>
 	<artifactId>lsq-parent</artifactId>
-	<version>2.0.0-RC2</version>
+	<version>2.0.0-RC3</version>
 
 	<packaging>pom</packaging>
 
@@ -63,19 +63,19 @@
 	</modules>
 
 	<properties>
-		<jena.version>4.8.0</jena.version>
-		<jenax.version>4.8.0-2</jenax.version>
-		<aksw-commons.version>0.9.6</aksw-commons.version>
-		<sansa.version>0.8.7</sansa.version>
+		<jena.version>5.0.0</jena.version>
+		<jenax.version>5.0.0-1</jenax.version>
+		<aksw-commons.version>0.9.8</aksw-commons.version>
+		<sansa.version>0.9.5</sansa.version>
 
-		<spark.version>3.0.1</spark.version>
+		<spark.version>3.5.1</spark.version>
 		<spark.deps.scope>compile</spark.deps.scope>
 
 		<scala.binary.version>2.12</scala.binary.version>
 		<scala.version.suffix>_${scala.binary.version}</scala.version.suffix>
 
-		<slf4j.version>1.7.36</slf4j.version>
-		<log4j.version>2.20.0</log4j.version>
+		<slf4j.version>2.0.9</slf4j.version>
+		<log4j.version>2.21.0</log4j.version>
 
 
 		<jackson.version>2.13.0</jackson.version>
@@ -87,7 +87,7 @@
 		<timestamp>${maven.build.timestamp}</timestamp>
 		<maven.build.timestamp.format>yyyy-MM-dd HH:mm</maven.build.timestamp.format>
 
-		<maven.compiler.release>11</maven.compiler.release>
+		<maven.compiler.release>17</maven.compiler.release>
 		<!-- <maven.compiler.source>1.8</maven.compiler.source> <maven.compiler.target>1.8</maven.compiler.target> -->
 		<java.version.javadoc>http://java.sun.com/javase/8/docs/api/</java.version.javadoc>
 		<file.encoding>UTF-8</file.encoding>
@@ -175,6 +175,12 @@
 				<version>4.4.0</version>
 			</dependency>
 
+			<dependency>
+				<groupId>com.google.guava</groupId>
+				<artifactId>guava</artifactId>
+				<version>32.1.3-jre</version>
+			</dependency>			
+
 			<!-- <dependency> -->
 			<!-- <groupId>net.sansa-stack</groupId> -->
 			<!-- <artifactId>sansa-rdf-spark${scala.version.suffix}</artifactId> -->
@@ -261,7 +267,7 @@
 			<dependency>
 				<groupId>org.aksw.thirdparty.org.spinrdf</groupId>
 				<artifactId>spinrdf</artifactId>
-				<version>2.0.0</version>
+				<version>2.0.0-jena5-1</version>
 				<exclusions>
 					<exclusion>
 						<groupId>org.apache.jena</groupId>
@@ -381,7 +387,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.8.1</version>
+				<version>3.11.0</version>
 				<configuration>
 					<release>${maven.compiler.release}</release>
 					<!-- <source>${maven.compiler.source}</source> <target>${maven.compiler.target}</target> -->
