@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Collections;
 import java.util.GregorianCalendar;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -23,14 +24,16 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.aksw.commons.io.util.StdIo;
 import org.aksw.jena_sparql_api.rx.query_flow.QueryFlowOps;
-import org.aksw.jenax.arq.connection.core.ConnectionLostException;
 import org.aksw.jenax.arq.dataset.api.ResourceInDataset;
+import org.aksw.jenax.arq.dataset.impl.DatasetGraphOneNgImpl;
+import org.aksw.jenax.arq.dataset.impl.DatasetOneNgImpl;
 import org.aksw.jenax.arq.dataset.impl.ResourceInDatasetImpl;
 import org.aksw.jenax.arq.util.binding.ResultSetUtils;
-import org.aksw.jenax.arq.util.exec.QueryExecutionUtils;
+import org.aksw.jenax.arq.util.exec.query.QueryExecutionUtils;
 import org.aksw.jenax.arq.util.expr.ExprUtils;
 import org.aksw.jenax.arq.util.quad.DatasetUtils;
 import org.aksw.jenax.arq.util.quad.Quads;
@@ -38,13 +41,14 @@ import org.aksw.jenax.arq.util.syntax.ElementUtils;
 import org.aksw.jenax.arq.util.syntax.QueryGenerationUtils;
 import org.aksw.jenax.arq.util.update.UpdateRequestUtils;
 import org.aksw.jenax.arq.util.var.Vars;
+import org.aksw.jenax.dataaccess.sparql.connection.reconnect.ConnectionLostException;
 import org.aksw.jenax.reprogen.core.MapperProxyUtils;
 import org.aksw.jenax.reprogen.hashid.HashIdCxt;
-import org.aksw.jenax.sparql.query.rx.RDFDataMgrRx;
 import org.aksw.jenax.sparql.query.rx.SparqlRx;
 import org.aksw.jenax.sparql.rx.op.FlowOfQuadsOps;
-import org.aksw.simba.lsq.enricher.core.LsqEnrichments;
+import org.aksw.simba.lsq.core.util.SkolemizeBackport;
 import org.aksw.simba.lsq.model.ExperimentConfig;
+import org.aksw.simba.lsq.model.ExperimentExec;
 import org.aksw.simba.lsq.model.ExperimentRun;
 import org.aksw.simba.lsq.model.LocalExecution;
 import org.aksw.simba.lsq.model.LsqQuery;
@@ -57,7 +61,6 @@ import org.aksw.simba.lsq.util.ElementVisitorFeatureExtractor;
 import org.aksw.simba.lsq.vocab.LSQ;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.jena.datatypes.xsd.XSDDateTime;
-import org.apache.jena.ext.com.google.common.base.Stopwatch;
 import org.apache.jena.graph.Node;
 import org.apache.jena.graph.NodeFactory;
 import org.apache.jena.query.Dataset;
@@ -73,11 +76,11 @@ import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.RDFNode;
 import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.rdfconnection.RDFConnection;
-import org.apache.jena.rdfconnection.RDFConnectionFactory;
 import org.apache.jena.rdfconnection.SparqlQueryConnection;
-import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.riot.RDFFormat;
+import org.apache.jena.riot.system.StreamRDF;
+import org.apache.jena.riot.system.StreamRDFOps;
 import org.apache.jena.sparql.core.DatasetGraphFactory;
 import org.apache.jena.sparql.core.Quad;
 import org.apache.jena.sparql.core.Var;
@@ -85,12 +88,14 @@ import org.apache.jena.sparql.engine.binding.Binding;
 import org.apache.jena.sparql.modify.request.QuadAcc;
 import org.apache.jena.sparql.syntax.ElementFilter;
 import org.apache.jena.sparql.syntax.Template;
+import org.apache.jena.sparql.util.ModelUtils;
 import org.apache.jena.system.Txn;
 import org.apache.jena.update.UpdateRequest;
-import org.apache.jena.util.ResourceUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spinrdf.model.TriplePattern;
+
+import com.google.common.base.Stopwatch;
 
 import io.reactivex.rxjava3.core.Flowable;
 import io.reactivex.rxjava3.core.FlowableTransformer;
@@ -111,42 +116,42 @@ public class LsqBenchmarkProcessor {
         return null;
     }
 
-
-    public static void run() {
-        SparqlQueryConnection benchmarkConn = RDFConnectionFactory.connect(DatasetFactory.create());
-
-        Model configModel = ModelFactory.createDefaultModel();
-
-        String expId = "testrun";
-        String expSuffix = "_" + expId;
-        String lsqBaseIri = "http://lsq.aksw.org/";
-
-        ExperimentConfig config = configModel
-                .createResource("http://someconfig.at/now")
-                .as(ExperimentConfig.class)
-                .setIdentifier(expId);
-
-        Instant benchmarkRunStartTimestamp = Instant.ofEpochMilli(0);
-        ZonedDateTime zdt = ZonedDateTime.ofInstant(benchmarkRunStartTimestamp, ZoneId.systemDefault());
-        Calendar cal = GregorianCalendar.from(zdt);
-        XSDDateTime xsddt = new XSDDateTime(cal);
-
-        ExperimentRun expRun = configModel
-                .createResource()
-                .as(ExperimentRun.class)
-                .setConfig(config)
-                .setTimestamp(xsddt);
-
-        HashIdCxt tmp = MapperProxyUtils.getHashId(expRun);
-        String expRunIri = lsqBaseIri + tmp.getStringId(expRun);
-        expRun = ResourceUtils.renameResource(expRun, expRunIri).as(ExperimentRun.class);
-
-        Flowable<LsqQuery> queryFlow = RDFDataMgrRx.createFlowableResources("../tmp/2020-06-27-wikidata-one-day.trig", Lang.TRIG, null)
-                .map(r -> r.as(LsqQuery.class));
-
-        // TODO Need to set up an index connection
-        process(queryFlow, lsqBaseIri, config, expRun, benchmarkConn, null);
-    }
+//
+//    public static void run() {
+//        SparqlQueryConnection benchmarkConn = RDFConnectionFactory.connect(DatasetFactory.create());
+//
+//        Model configModel = ModelFactory.createDefaultModel();
+//
+//        String expId = "testrun";
+//        String expSuffix = "_" + expId;
+//        String lsqBaseIri = "http://lsq.aksw.org/";
+//
+//        ExperimentConfig config = configModel
+//                .createResource("http://someconfig.at/now")
+//                .as(ExperimentConfig.class)
+//                .setIdentifier(expId);
+//
+//        Instant benchmarkRunStartTimestamp = Instant.ofEpochMilli(0);
+//        ZonedDateTime zdt = ZonedDateTime.ofInstant(benchmarkRunStartTimestamp, ZoneId.systemDefault());
+//        Calendar cal = GregorianCalendar.from(zdt);
+//        XSDDateTime xsddt = new XSDDateTime(cal);
+//
+//        ExperimentRun expRun = configModel
+//                .createResource()
+//                .as(ExperimentRun.class)
+//                .setConfig(config)
+//                .setTimestamp(xsddt);
+//
+//        HashIdCxt tmp = MapperProxyUtils.getHashId(expRun);
+//        String expRunIri = lsqBaseIri + tmp.getStringId(expRun);
+//        expRun = ResourceUtils.renameResource(expRun, expRunIri).as(ExperimentRun.class);
+//
+//        Flowable<LsqQuery> queryFlow = RDFDataMgrRx.createFlowableResources("../tmp/2020-06-27-wikidata-one-day.trig", Lang.TRIG, null)
+//                .map(r -> r.as(LsqQuery.class));
+//
+//        // TODO Need to set up an index connection
+//        process(queryFlow, lsqBaseIri, config, expRun, benchmarkConn, null);
+//    }
 
     /**
      * If something goes wrong when running the wrapped action
@@ -167,6 +172,25 @@ public class LsqBenchmarkProcessor {
        return result;
    }
 
+   public static void process(
+           StreamRDF out,
+           Flowable<LsqQuery> rawQueryFlow,
+           String lsqBaseIri,
+           ExperimentConfig expConfig,
+           ExperimentExec expExec,
+           ExperimentRun expRun,
+           Function<Resource, Resource> enricher,
+           SparqlQueryConnection benchmarkConn,
+           RDFConnection indexConn) {
+       Flowable<ResourceInDataset> flowable = processCore(rawQueryFlow, lsqBaseIri, expConfig, expExec, expRun, enricher, benchmarkConn, indexConn);
+
+       Iterable<ResourceInDataset> items = flowable.blockingIterable();
+       for(ResourceInDataset item : items) {
+           // RDFDataMgr.write(out, item.getDataset(), RDFFormat.TRIG_BLOCKS);
+           StreamRDFOps.sendDatasetToStream(item.getDataset().asDatasetGraph(), out);
+       }
+   }
+
     /**
      *
      * @param lsqBaseIri The IRI prefix to prepend to generated resources
@@ -175,11 +199,13 @@ public class LsqBenchmarkProcessor {
      *        of benchmark information
      * @param benchmarkConn The connection on which to perform benchmarking
      */
-    public static void process(
+    public static Flowable<ResourceInDataset> processCore(
             Flowable<LsqQuery> rawQueryFlow,
             String lsqBaseIri,
-            ExperimentConfig config,
+            ExperimentConfig expConfig,
+            ExperimentExec expExec,
             ExperimentRun expRun,
+            Function<Resource, Resource> enricher,
             SparqlQueryConnection benchmarkConn,
             RDFConnection indexConn) {
 
@@ -188,15 +214,16 @@ public class LsqBenchmarkProcessor {
         //String expSuffix
 
 
-        String datasetLabel = config.getDatasetLabel();
-        XSDDateTime benchmarkRunTimestamp = expRun.getTimestamp();
-        Instant instant = benchmarkRunTimestamp.asCalendar().toInstant();
+        String datasetLabel = expConfig.getDatasetLabel();
+        XSDDateTime benchmarkExecTimestamp = expExec.getTimestamp();
+        Instant instant = benchmarkExecTimestamp.asCalendar().toInstant();
+        String runId = "run" + (Optional.ofNullable(expRun.getRunId()).map(Object::toString).orElse(""));
         ZonedDateTime zdt = ZonedDateTime.ofInstant(instant, ZoneId.systemDefault());
-        String benchmarkRunTimestampStr = DateTimeFormatter.ISO_LOCAL_DATE.format(zdt);
-        String expId = datasetLabel + "_" + benchmarkRunTimestampStr;
-        String expSuffix = "_" + expId;
+        String benchmarkExecTimestampStr = DateTimeFormatter.ISO_LOCAL_DATE.format(zdt);
+        String expId = datasetLabel + "_" + benchmarkExecTimestampStr;
+        String expSuffix = "_" + expId + "_" + runId;
 
-        boolean benchmarkSecondaryQueries = Optional.ofNullable(config.benchmarkSecondaryQueries()).orElse(false);
+        boolean benchmarkSecondaryQueries = Optional.ofNullable(expConfig.benchmarkSecondaryQueries()).orElse(false);
 
         Function<String, String> lsqQueryBaseIriFn = hash -> lsqBaseIri + "q-" + hash;
 
@@ -211,6 +238,11 @@ public class LsqBenchmarkProcessor {
 //                .map(r -> r.as(LsqQuery.class))
 //                .skip(1)
 //                .take(1)
+                .concatMapMaybe(lsqQuery -> {
+                    Maybe<LsqQuery> r = safeMaybe(() -> enricher.apply(lsqQuery).as(LsqQuery.class));
+                    return r;
+                })
+                /*
                 .concatMapMaybe(lsqQuery ->
                     safeMaybe(() -> LsqEnrichments.enrichWithFullSpinModelCore(lsqQuery)))
 //                .concatMapMaybe(lsqQuery -> enrichWithFullSpinModel(lsqQuery))
@@ -221,9 +253,11 @@ public class LsqBenchmarkProcessor {
 //                    RDFDataMgr.write(System.out, x.getModel(), RDFFormat.TURTLE_FLAT);
 //                    System.exit(1);
 //                })
+   */
                 //.flatMap(lsqQuery -> Flowable.fromIterable(extractAllQueries(lsqQuery)), false, 128)
-                .map(lsqQuery -> extractAllQueries(lsqQuery))
-                .map(batch -> benchmarkSecondaryQueries ? batch : Collections.singleton(batch.iterator().next()))
+                //.map(lsqQuery -> extractAllQueries(lsqQuery))
+                //.map(batch -> benchmarkSecondaryQueries ? batch : Collections.singleton(batch.iterator().next()))
+                .map(lsqQuery -> benchmarkSecondaryQueries ? extractAllQueries(lsqQuery) : Collections.singleton(lsqQuery))
 //                .doAfterNext(lsqQuery -> lsqQuery.updateHash())
 //                .doOnNext(r -> ResourceUtils.renameResource(r, "http://lsq.aksw.org/q-" + r.getHash()).as(LsqQuery.class))
 //                .lift(OperatorObserveThroughput.create("throughput", 100))
@@ -231,25 +265,42 @@ public class LsqBenchmarkProcessor {
 //                .lift(OperatorObserveThroughput.create("buffered", 100))
                 ;
 
-        Iterable<List<Set<LsqQuery>>> batches = queryFlow.blockingIterable();
-        Iterator<List<Set<LsqQuery>>> itBatches = batches.iterator();
-
-        // Create a database to ensure uniqueness of evaluation tasks
-        while(itBatches.hasNext()) {
-            List<Set<LsqQuery>> batch = itBatches.next();
+        Flowable<ResourceInDataset> result = queryFlow.flatMapIterable(batch -> {
             List<ResourceInDataset> items = processBatchOfQueries(
                     batch,
                     lsqBaseIri,
-                    config,
+                    expConfig,
+                    expExec,
                     expRun,
                     benchmarkConn,
                     lsqQueryExecFn,
                     indexConn);
+            return items;
+        });
 
-            for(ResourceInDataset item : items) {
-                RDFDataMgr.write(StdIo.openStdOutWithCloseShield(), item.getDataset(), RDFFormat.TRIG_BLOCKS);
+        if (false) {
+            Iterable<List<Set<LsqQuery>>> batches = queryFlow.blockingIterable();
+            Iterator<List<Set<LsqQuery>>> itBatches = batches.iterator();
+
+            // Create a database to ensure uniqueness of evaluation tasks
+            while(itBatches.hasNext()) {
+                List<Set<LsqQuery>> batch = itBatches.next();
+                List<ResourceInDataset> items = processBatchOfQueries(
+                        batch,
+                        lsqBaseIri,
+                        expConfig,
+                        expExec,
+                        expRun,
+                        benchmarkConn,
+                        lsqQueryExecFn,
+                        indexConn);
+
+                for(ResourceInDataset item : items) {
+                    RDFDataMgr.write(StdIo.openStdOutWithCloseShield(), item.getDataset(), RDFFormat.TRIG_BLOCKS);
+                }
             }
         }
+        return result;
 
 
 //
@@ -307,149 +358,141 @@ public class LsqBenchmarkProcessor {
 //        RDFDataMgr.write(System.out, model, RDFFormat.TURTLE_PRETTY);
     }
 
+    /** Create a union model of all unique non-null arguments */
+    public static Model unionAll(Model ...models) {
+        Model result = Stream.of(models)
+                .filter(Objects::nonNull)
+                .distinct()
+                .collect(ModelUtils.unionCollector());
+        return result;
+    }
 
     public static List<ResourceInDataset> processBatchOfQueries(
             List<Set<LsqQuery>> batch,
             String lsqBaseIri,
-            ExperimentConfig config,
+            ExperimentConfig expConfig,
+            ExperimentExec expExec,
             ExperimentRun expRun,
             SparqlQueryConnection benchmarkConn,
             Function<LsqQuery, String> lsqQueryExecFn,
             RDFConnection indexConn) {
 
+        // Skolemization is blocked for all resources appearing in the unionModel
+        Model staticModel = unionAll(expConfig.getModel(), expExec.getModel(), expRun.getModel());
+
+        // Combine the query hash and the exprRun id to form the benchmark task id.
         List<ResourceInDataset> result = new ArrayList<>();
 
-        Set<Node> lookupHashNodes = batch.stream()
-                .flatMap(item -> item.stream())
-                .map(lsqQueryExecFn)
-                .map(NodeFactory::createURI)
-                .collect(Collectors.toSet());
+        Map<Node, LsqQuery> inputTasks = new HashMap<>();
+        for (Set<LsqQuery> queries : batch) {
+            for (LsqQuery query : queries) {
+                String taskIdStr = lsqQueryExecFn.apply(query);
+                Node taskId = NodeFactory.createURI(taskIdStr);
+                if (inputTasks.containsKey(taskId)) {
+                    throw new RuntimeException("Task already exists: " + taskId);
+                }
+                inputTasks.put(taskId, query);
+            }
+        }
 
-//                for(Node n : lookupHashNodes) {
-//                    System.out.println("Lookup with: " + n.getURI());
-//                }
-//                Expr filter = ExprUtils.oneOf(Vars.g, lookupHashNodes);
-//                dq.filterDirect(new ElementFilter(filter));
-
-        Map<String, Dataset> nodeToDataset = Txn.calculate(indexConn, () ->
-            fetchDatasets(indexConn, lookupHashNodes)
+        Map<String, Dataset> taskIdToDataset = Txn.calculate(indexConn, () ->
+            fetchDatasets(indexConn, inputTasks.keySet())
             .toMap(Entry::getKey, Entry::getValue)
             .blockingGet());
 
-//                System.out.println(hashNodes);
         // Obtain the set of query strings already in the store
+        Set<String> completedTaskIds = taskIdToDataset.keySet();
 
-        Set<String> alreadyIndexedHashUrns = nodeToDataset.keySet();
-//                Set<String> alreadyIndexedHashUrns = Txn.calculate(indexConn, () ->
-//                      dq
-//                        .exec()
-//                        .map(x -> x.asNode().getURI())
-//                        .blockingStream()
-//                        .collect(Collectors.toSet()));
-//                        ;
+        Map<Node, LsqQuery> pendingTasks = inputTasks.entrySet().stream()// batch.stream()
+            .filter(e -> !completedTaskIds.contains(e.getKey().toString()))
+            .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
 
-//                for(String str : alreadyIndexedHashUrns) {
-//                    System.out.println("Already indexed: " + str);
-//                }
-        Set<LsqQuery> nonIndexed = batch.stream()
-            .flatMap(item -> item.stream())
-            .filter(item -> !alreadyIndexedHashUrns.contains(lsqQueryExecFn.apply(item)))
-            .collect(Collectors.toSet());
+        List<Quad> inserts = new ArrayList<>();
 
-//                System.out.println(nonIndexed);
+        for (Entry<Node, LsqQuery> task : pendingTasks.entrySet()) {
+            Node queryExecId = task.getKey();
+            String queryExecIri = queryExecId.getURI();
 
-        List<Quad> inserts = new ArrayList<Quad>();
+            LsqQuery lsqQuery = task.getValue();
+            String queryStr = lsqQuery.getText();
 
-//                System.err.println("Batch: " + nonIndexed.size() + "/" + lookupHashNodes.size() + " (" + batch.size() + ") need processing");
-        for(LsqQuery item : nonIndexed) {
-            String queryStr = item.getText();
-            String queryExecIri = lsqQueryExecFn.apply(item); //"urn://" + item.getHash();
-            Node s = NodeFactory.createURI(queryExecIri);
-//                    System.out.println("Benchmarking: " + s);
-//                    System.out.println(item.getText());
-
-            Dataset newDataset = DatasetFactory.create();
-            Model newModel = newDataset.getNamedModel(queryExecIri);
-
-            item = item.inModel(newModel).as(LsqQuery.class);
+            Model newModel = ModelFactory.createDefaultModel();
+            LsqQuery newLsqQuery = lsqQuery.inModel(newModel).as(LsqQuery.class);
 
             // Create fresh local execution and query exec resources
-            LocalExecution le = newModel.createResource().as(LocalExecution.class);
-            QueryExec qe = newModel.createResource().as(QueryExec.class);
+            LocalExecution newLocalExec = newModel.createResource().as(LocalExecution.class);
+            QueryExec newQueryExec = newModel.createResource().as(QueryExec.class);
 
-            // TODO Skolemize these executions! Update: Its done - isn't it/
             rdfizeQueryExecutionBenchmark(
                     benchmarkConn,
                     queryStr,
-                    qe,
-                    config.getConnectionTimeoutForRetrieval(),
-                    config.getExecutionTimeoutForRetrieval(),
-                    config.getMaxResultCountForCounting(),
-                    config.getMaxByteSizeForCounting(),
-                    config.getMaxResultCountForSerialization(),
-                    config.getMaxByteSizeForSerialization(),
-                    config.getConnectionTimeoutForCounting(),
-                    config.getConnectionTimeoutForRetrieval(),
-                    config.getMaxCount(),
-                    config.getMaxCountAffectsTp());
+                    newQueryExec,
+                    expConfig.getConnectionTimeoutForRetrieval(),
+                    expConfig.getExecutionTimeoutForRetrieval(),
+                    expConfig.getMaxResultCountForCounting(),
+                    expConfig.getMaxByteSizeForCounting(),
+                    expConfig.getMaxResultCountForSerialization(),
+                    expConfig.getMaxByteSizeForSerialization(),
+                    expConfig.getConnectionTimeoutForCounting(),
+                    expConfig.getConnectionTimeoutForRetrieval(),
+                    expConfig.getMaxCount(),
+                    expConfig.getMaxCountAffectsTp());
 
-            item.getLocalExecutions().add(le);
-            le.setBenchmarkRun(expRun);
-            le.setQueryExec(qe);
+            newLsqQuery.getLocalExecutions().add(newLocalExec);
+            newLocalExec.setBenchmarkRun(expRun);
+            newLocalExec.setQueryExec(newQueryExec);
 
-            newDataset.asDatasetGraph().find().forEachRemaining(inserts::add);
+            // Skolemization is blocked for all resources appearing in the unionModel
+            Model unionModel = unionAll(staticModel, lsqQuery.getModel());
 
+            // So we only skolemize all resources related to the newLocalExec
+            LocalExecution finalLocalExec = SkolemizeBackport.skolemize(newLocalExec, unionModel, lsqBaseIri, LocalExecution.class);
 
-            inserts.add(new Quad(s, s, LSQ.execStatus.asNode(), NodeFactory.createLiteral("processed")));
-//                    RDFDataMgr.write(new FileOutputStream(FileDescriptor.out), item.getModel(), RDFFormat.TURTLE_PRETTY);
-
-//                    System.out.println(s.getURI().equals("urn://33c4205f90fdeb7d1db68426806a816e3ffbfa3980461b556b32fded407568c9"));
-
-            nodeToDataset.put(queryExecIri, newDataset);
+            Dataset newDataset = new DatasetOneNgImpl(DatasetGraphOneNgImpl.create(queryExecId, finalLocalExec.getModel().getGraph()));
+            inserts.add(new Quad(queryExecId, queryExecId, LSQ.execStatus.asNode(), NodeFactory.createLiteralString("processed")));
+            taskIdToDataset.put(queryExecIri, newDataset);
         }
 
         UpdateRequest ur = UpdateRequestUtils.createUpdateRequest(inserts, null);
         Txn.executeWrite(indexConn, () -> indexConn.update(ur));
 
-
         // Remove the execStatus "processed" triples from the fetched datasets
-        for(Dataset ds : nodeToDataset.values()) {
+        for(Dataset ds : taskIdToDataset.values()) {
             for(Entry<String, Model> e : DatasetUtils.listModels(ds)) {
                 e.getValue().removeAll(null, LSQ.execStatus, null);
             }
         }
-//                Txn.executeRead(indexConn, () -> System.out.println(ResultSetFormatter.asText(indexConn.query("SELECT ?s { ?s ?p ?o }").execSelect())));
+
+        // Txn.executeRead(indexConn, () -> System.out.println(ResultSetFormatter.asText(indexConn.query("SELECT ?s { ?s ?p ?o }").execSelect())));
 
         for(Set<LsqQuery> pack : batch) {
 
             logger.info("Processing pack of size: " + pack.size());
-// TODO Move all the code into a nice processPack method of a new class
-try {
-            // The primary query is assumed to always be the first element of a pack
-            LsqQuery primaryQuery = pack.iterator().next();
+            // TODO Move all the code into a nice processPack method of a new class
+            try {
+                // The primary query is assumed to always be the first element of a pack
+                LsqQuery primaryQueryRaw = pack.iterator().next();
 
 
-            Model primaryQueryModel = ModelFactory.createDefaultModel();
+                Model primaryQueryModel = ModelFactory.createDefaultModel();
+
+                // We need to add the config model in order to include the benchmark run id
+                // We remove the config once we are done
+
+                // TODO We should ensure that only the minimal necessary config model is added
+                // expRoot.getModel().add(configModel);
+
+//                primaryQueryModel.add(configModel);
+//                primaryQueryModel.add(expRun.getModel());
 
 
-            // We need to add the config model in order to include the benchmark run id
-            // We remove the config once we are done
-
-            // TODO We should ensure that only the minimal necessary config model is added
-            Model configModel = config.getModel();
-//                    expRoot.getModel().add(configModel);
-            primaryQueryModel.add(configModel);
-
-
-            //Model model = rootQuery.getModel();
-
-            // Extend the rootQuery's model with all related query executions
-            for(LsqQuery item : pack) {
-                String key = lsqQueryExecFn.apply(item);
-                Dataset ds = nodeToDataset.get(key);
-                Objects.requireNonNull(ds, "Expected dataset for key "  + key);
-                Model m = ds.getNamedModel(key);
-                Objects.requireNonNull(m, "Should not happen: No query execution model for " + key);
+                // Extend the rootQuery's model with all related query executions
+                for(LsqQuery item : pack) {
+                    String key = lsqQueryExecFn.apply(item);
+                    Dataset ds = taskIdToDataset.get(key);
+                    Objects.requireNonNull(ds, "Expected dataset for key "  + key);
+                    Model m = ds.getNamedModel(key);
+                    Objects.requireNonNull(m, "Should not happen: No query execution model for " + key);
 
 //                        System.err.println("BEGIN***********************************************");
 //                        if(item.getModel().contains(ResourceFactory.createResource("http://www.bigdata.com/rdf#serviceParam"), null, (RDFNode)null)) {
@@ -458,65 +501,76 @@ try {
 //                        RDFDataMgr.write(System.err, item.getModel(), RDFFormat.TURTLE_PRETTY);
 //                        System.err.println("END***********************************************");
 
-                // Adding the primary query's model to itself should be harmless
-                primaryQueryModel.add(m);
-                primaryQueryModel.add(item.getModel());
+                    // Adding the primary query's model to itself should be harmless
+                    primaryQueryModel.add(m);
+                    primaryQueryModel.add(item.getModel());
+                }
+
+
+                LsqQuery primaryQuery = primaryQueryRaw.inModel(primaryQueryModel).as(LsqQuery.class);
+
+                // Update triple pattern selectivities
+                // LocalExecution expRoot = model.createResource().as(LocalExecution.class);
+                Map<Resource, LocalExecution> rleMap = primaryQuery.getLocalExecutionMap();
+                LocalExecution expRoot = rleMap.get(expRun);
+
+                if (expRoot == null) {
+                    throw new RuntimeException("Should not happen: No local execution with id " + expRun.asNode() +
+                            ". Candidates: " + rleMap.keySet().stream().map(RDFNode::asNode).collect(Collectors.toSet()));
+                }
+
+                // expRoot.setBenchmarkRun(expRun);
+
+                // If there is no spin model then don't try to create executions for its elements
+                if (primaryQuery.getSpinQuery() != null) {
+                    LsqExec.createAllExecs(primaryQuery, expRun);
+                }
+
+                if (false) {
+                    Model configModel = expConfig.getModel();
+
+                    HashIdCxt hashIdCxt = MapperProxyUtils.getHashId(expRoot);//.getHash(bgp);
+                    //Map<RDFNode, HashCode> renames = hashIdCxt.getMapping();
+                    Map<RDFNode, String> renames = hashIdCxt.getStringIdMapping();
+
+        //                    Map<Resource, String> renames = new LinkedHashMap<>();
+        //                    for(SpinBgp bgp : spinRoot.getBgps()) {
+        //                        HashIdCxt hashIdCxt = MapperProxyUtils.getHashId(bgp);//.getHash(bgp);
+        //
+        //                        for(Entry<RDFNode, HashCode> e : hashIdCxt.getMapping().entrySet()) {
+        //                            if(e.getKey().isResource()) {
+        //                                renames.put(e.getKey().asResource(), e.getValue().toString());
+        //                            }
+        //                        }
+        //                    }
+
+        //                    for(Entry<RDFNode, HashCode> e : renames.entrySet()) {
+
+                    primaryQueryModel.remove(expRun.getModel());
+                    primaryQueryModel.remove(configModel);
+
+                    Map<Resource, Resource> remap = org.aksw.jena_sparql_api.rdf.collections.ResourceUtils.renameResources(lsqBaseIri, renames);
+
+
+                    // If the primaryQuery was renamed
+                    Resource tgtPrimaryQuery = remap.getOrDefault(primaryQuery, primaryQuery);
+                }
+
+                // primaryQueryModel.remove(expRun.getModel());
+                // primaryQueryModel.remove(configModel);
+
+                Resource tgtPrimaryQuery = primaryQuery;
+                //String graphIri = primaryQuery.getURI();
+                ResourceInDataset item = ResourceInDatasetImpl.createFromCopyIntoResourceGraph(tgtPrimaryQuery);
+                result.add(item);
+            } catch (Exception e) {
+                logger.warn("Internal error; trying to continue", e);
             }
-
-
-            primaryQuery = primaryQuery.inModel(primaryQueryModel).as(LsqQuery.class);
-
-            // Update triple pattern selectivities
-//                    LocalExecution expRoot = model.createResource().as(LocalExecution.class);
-            Map<Resource, LocalExecution> rleMap = primaryQuery.getLocalExecutionMap();
-            LocalExecution expRoot = rleMap.get(expRun);
-
-//                    expRoot.setBenchmarkRun(expRun);
-
-            LsqExec.createAllExecs(primaryQuery, expRun);
-
-
-
-            HashIdCxt hashIdCxt = MapperProxyUtils.getHashId(expRoot);//.getHash(bgp);
-            //Map<RDFNode, HashCode> renames = hashIdCxt.getMapping();
-            Map<RDFNode, String> renames = hashIdCxt.getStringIdMapping();
-
-//                    Map<Resource, String> renames = new LinkedHashMap<>();
-//                    for(SpinBgp bgp : spinRoot.getBgps()) {
-//                        HashIdCxt hashIdCxt = MapperProxyUtils.getHashId(bgp);//.getHash(bgp);
-//
-//                        for(Entry<RDFNode, HashCode> e : hashIdCxt.getMapping().entrySet()) {
-//                            if(e.getKey().isResource()) {
-//                                renames.put(e.getKey().asResource(), e.getValue().toString());
-//                            }
-//                        }
-//                    }
-
-//                    for(Entry<RDFNode, HashCode> e : renames.entrySet()) {
-
-            primaryQueryModel.remove(configModel);
-
-            Map<Resource, Resource> remap = org.aksw.jena_sparql_api.rdf.collections.ResourceUtils.renameResources(lsqBaseIri, renames);
-
-
-            // If the primaryQuery was renamed
-            Resource tgtPrimaryQuery = remap.getOrDefault(primaryQuery, primaryQuery);
-
-            //String graphIri = primaryQuery.getURI();
-            ResourceInDataset item = ResourceInDatasetImpl.createFromCopyIntoResourceGraph(tgtPrimaryQuery);
-            result.add(item);
-}
-catch (Exception e) {
-    logger.warn("Internal error; trying to continue", e);
-}
-
             //RDFDataMgr.write(System.out, spinRoot.getModel(), RDFFormat.TURTLE_BLOCKS);
         }
 
         return result;
     }
-
-
 
     public static LsqQuery updateLsqQueryIris(
             LsqQuery start,
@@ -688,6 +742,7 @@ catch (Exception e) {
                return result;
            }
 
+           // TODO For COUNT queries add the LSQ.countValue to the output model
 
            Stopwatch evalSw = Stopwatch.createStarted(); // Total time spent evaluating
 
@@ -701,10 +756,13 @@ catch (Exception e) {
                logger.info("Benchmarking " + queryStr);
                Stopwatch retrievalSw = Stopwatch.createStarted();
 
-               try(QueryExecution qe = conn.query(query)) {
+               try(QueryExecution qe = conn.newQuery()
+                       .timeout(executionTimeoutForRetrieval, TimeUnit.MILLISECONDS)
+                       .query(query)
+                       .build()) {
                    // https://github.com/apache/jena/issues/1384
                    // qe.setTimeout(connectionTimeoutForRetrieval, executionTimeoutForRetrieval);
-                   qe.setTimeout(executionTimeoutForRetrieval);
+                   // qe.setTimeout(executionTimeoutForRetrieval);
 
                    ResultSet rs = qe.execSelect();
                    varNames.addAll(rs.getResultVars());
@@ -823,14 +881,19 @@ catch (Exception e) {
                    Var countVar = queryAndVar.getKey();
                    Query countQuery = queryAndVar.getValue();
 
-                   logger.info("Counting " + countQuery);
+                   if (logger.isInfoEnabled()) {
+                       logger.info("Counting " + countQuery);
+                   }
 
                    countingSw = Stopwatch.createStarted();
 
-                   try(QueryExecution qe = conn.query(countQuery)) {
+                   try(QueryExecution qe = conn.newQuery()
+                           .query(countQuery)
+                           .timeout(executionTimeoutForCounting, TimeUnit.MILLISECONDS)
+                           .build()) {
                        // qe.setTimeout(connectionTimeoutForCounting, executionTimeoutForCounting);
                        // https://github.com/apache/jena/issues/1384
-                       qe.setTimeout(executionTimeoutForCounting);
+                       // qe.setTimeout(executionTimeoutForCounting);
                        Number count = QueryExecutionUtils.fetchNumber(qe, countVar);
                        if(count != null) {
                            itemCount = count.longValue();
@@ -841,7 +904,9 @@ catch (Exception e) {
                } catch (ConnectionLostException e) {
                    throw new ConnectionLostException(e);
                } catch(Exception e) {
-                   logger.warn("Counting error: ", e);
+                   if (logger.isWarnEnabled()) {
+                       logger.warn("Counting error: ", e);
+                   }
     //                   String errorMsg = Optional.ofNullable(ExceptionUtils.getRootCause(e)).orElse(e).getMessage();
                    String errorMsg = ExceptionUtils.getStackTrace(e);
                    result.setCountingError(errorMsg);
@@ -873,7 +938,11 @@ catch (Exception e) {
 
            result.setEvalDuration(evalDuration);
 
-           logger.info("Benchmark result after " + evalDuration + " seconds: " + result.getResultSetSize() + " results and error message " + result.getRetrievalError());
+           if(logger.isInfoEnabled()) {
+               String errMsg = result.getRetrievalError();
+               logger.info("Benchmark result after " + evalDuration + " seconds: " + result.getResultSetSize() + " results"
+                       + (errMsg == null ? " (success)" : " and error message: " + errMsg));
+           }
 
            return result;
            //Calendar end = Calendar.getInstance();
