@@ -1,9 +1,8 @@
 package org.aksw.simba.lsq.util;
 
-import org.aksw.jena_sparql_api.concepts.Concept;
-import org.aksw.jena_sparql_api.concepts.ConceptUtils;
+import org.aksw.jenax.sparql.fragment.api.Fragment1;
+import org.aksw.jenax.sparql.fragment.impl.ConceptUtils;
 import org.aksw.jenax.sparql.query.rx.SparqlRx;
-import org.aksw.jenax.sparql.relation.api.UnaryRelation;
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryExecutionFactory;
 import org.apache.jena.rdf.model.Model;
@@ -14,7 +13,7 @@ import org.apache.jena.sparql.core.Var;
 import io.reactivex.rxjava3.core.Flowable;
 
 public class ConceptModelUtils {
-    public static <T extends Resource> Flowable<T> listResources(Model model, Concept concept, Class<? extends T> clazz) {
+    public static <T extends Resource> Flowable<T> listResources(Model model, Fragment1 concept, Class<? extends T> clazz) {
         return listRdfNodes(model, concept)
 //                .filter(rdfNode -> rdfNode.canAs(clazz))
                 .filter(rdfNode -> {
@@ -43,7 +42,7 @@ public class ConceptModelUtils {
      */
     public static <T extends Resource> Flowable<T> listResourcesUnchecked(
             Model model,
-            UnaryRelation concept,
+            Fragment1 concept,
             Class<? extends T> clazz) {
         return listRdfNodes(model, concept)
                 .map(rdfNode -> {
@@ -63,7 +62,7 @@ public class ConceptModelUtils {
 //        return result;
 //    }
 
-    public static Flowable<RDFNode> listRdfNodes(Model model, UnaryRelation concept) {
+    public static Flowable<RDFNode> listRdfNodes(Model model, Fragment1 concept) {
         Var var = concept.getVar();
         Query query = ConceptUtils.createQueryList(concept);
 
