@@ -14,8 +14,8 @@ import org.aksw.simba.lsq.parser.StringMapper;
 import org.aksw.simba.lsq.parser.WebLogParser;
 import org.aksw.simba.lsq.vocab.LSQ;
 import org.aksw.simba.lsq.vocab.PROV;
-import org.apache.jena.ext.com.google.common.reflect.ClassPath;
-import org.apache.jena.ext.com.google.common.reflect.ClassPath.ResourceInfo;
+import com.google.common.reflect.ClassPath;
+import com.google.common.reflect.ClassPath.ResourceInfo;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.riot.RDFDataMgr;
