@@ -17,8 +17,8 @@ import org.aksw.simba.lsq.core.LsqRdfizer;
 import org.aksw.simba.lsq.core.LsqUtils;
 import org.apache.commons.compress.compressors.bzip2.BZip2CompressorInputStream;
 import org.apache.jena.atlas.iterator.Iter;
-import org.apache.jena.ext.com.google.common.base.Stopwatch;
-import org.apache.jena.ext.com.google.common.collect.Iterators;
+import com.google.common.base.Stopwatch;
+import com.google.common.collect.Iterators;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.query.Syntax;
 import org.apache.jena.rdf.model.Model;
