@@ -39,8 +39,8 @@ import org.aksw.simba.lsq.vocab.LSQ;
 import org.aksw.simba.lsq.vocab.PROV;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.jena.datatypes.xsd.XSDDateTime;
-import org.apache.jena.ext.com.google.common.base.Stopwatch;
-import org.apache.jena.ext.com.google.common.hash.Hashing;
+import com.google.common.base.Stopwatch;
+import com.google.common.hash.Hashing;
 import org.apache.jena.graph.Node;
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryExecution;
