@@ -6,8 +6,8 @@ import java.util.Map.Entry;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
-import org.aksw.commons.util.stream.SequentialGroupBySpec;
-import org.aksw.commons.util.stream.StreamOperatorSequentialGroupBy;
+import org.aksw.commons.util.stream.CollapseRunsSpec;
+import org.aksw.commons.util.stream.StreamOperatorCollapseRuns;
 import org.apache.jena.graph.Node;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.query.DatasetFactory;
@@ -17,7 +17,7 @@ import org.apache.jena.sparql.core.Quad;
 public class DatasetGraphOpsStream {
     public static StreamTransformer<Quad, Entry<Node, List<Quad>>> groupToList()
     {
-        return StreamOperatorSequentialGroupBy.<Quad, Node, List<Quad>>create(SequentialGroupBySpec.create(
+        return StreamOperatorCollapseRuns.<Quad, Node, List<Quad>>create(CollapseRunsSpec.create(
                 Quad::getGraph,
                 graph -> new ArrayList<>(),
                 (list, item) -> list.add(item)
@@ -28,7 +28,7 @@ public class DatasetGraphOpsStream {
             Function<Quad, Node> grouper,
             Supplier<? extends DatasetGraph> graphSupplier) {
 
-        return StreamOperatorSequentialGroupBy.<Quad, Node, DatasetGraph>create(SequentialGroupBySpec.create(
+        return StreamOperatorCollapseRuns.<Quad, Node, DatasetGraph>create(CollapseRunsSpec.create(
                 grouper::apply,
                 groupKey -> graphSupplier.get(),
                 DatasetGraph::add))::transform;
